Split out files and add test project
