+++ src/Orleans.Testing.Utils/Properties/AssemblyInfo.cs
@@ -1,0 +1,3 @@
+﻿using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Orleans.Testing.Utils.Tests", AllInternalsVisible = true)]
